# catalog-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API server for a catalog service.",
        "title": "Catalog Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3003",
    "basePath": "/",
    "paths": {
...
        "/api/v1/catalog/menus/{id}/portions": {
            "get": {
                "description": "Get size variants for a specific menu item",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "portions"
                ],
                "summary": "Get portions by Menu Item ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Menu Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/model.MenuPortionSize"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new size variant for a menu item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "portions"
                ],
                "summary": "Create a portion size",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Menu Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Portion Data",
                        "name": "portion",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.MenuPortionSize"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.MenuPortionSize"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/v1/catalog/portions/{portion_id}": {
            "delete": {
                "description": "Delete a size variant by its ID",
                "tags": [
                    "portions"
                ],
                "summary": "Delete a portion size",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Portion ID",
                        "name": "portion_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.MenuPortionSize": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "extra_price": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "is_default": {
                    "type": "boolean"
                },
                "menu_item_id": {
                    "type": "integer"
                },
                "name": {
                    "description": "e.g. \"ธรรมดา\", \"พิเศษ\"",
                    "type": "string"
                },
                "quantity_multiplier": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    }
}